# packages/frontend/src/redux/schema/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/{id}": {
      "get": {
        "operationId": "Users_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Entity id from users collection.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          }
        },
        "tags": ["Users"]
      },
      "patch": {
        "operationId": "Users_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Entity id from users collection.",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDto" }
              }
            }
          }
        },
        "tags": ["Users"]
      }
    },
...
    "/api/user-preferences": {
      "get": {
        "operationId": "UserPreferences_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserPreferenceDto" }
                }
              }
            }
          }
        },
        "tags": ["User preferences"]
      },
      "patch": {
        "operationId": "UserPreferences_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserPreferenceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserPreferenceDto" }
              }
            }
          }
        },
        "tags": ["User preferences"]
      }
    },
    "/api/user-preferences/{userPreferenceProperty}": {
      "get": {
        "operationId": "UserPreferences_findOne",
        "parameters": [
          {
            "name": "userPreferenceProperty",
            "required": true,
            "in": "path",
            "schema": { "$ref": "#/components/schemas/UserPreferenceProperty" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserPreferenceDto" }
              }
            }
          }
        },
        "tags": ["User preferences"]
      }
    },
...
  },
  "info": {
    "title": "Financer",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "Role": { "type": "string", "enum": ["ADMIN", "TEST_USER"] },
      "Theme": { "type": "string", "enum": ["AUTO", "LIGHT", "DARK"] },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "name": { "type": "string" },
          "nickname": { "type": "string" },
          "githubId": { "type": "string", "nullable": true },
          "auth0Id": { "type": "string", "nullable": true },
          "profileImageUrl": { "type": "string", "nullable": true },
          "roles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Role" }
          },
          "theme": { "allOf": [{ "$ref": "#/components/schemas/Theme" }] }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "nickname",
          "githubId",
          "auth0Id",
          "profileImageUrl",
          "roles",
          "theme"
        ]
      },
...
      "UserPreferenceProperty": {
        "type": "string",
        "enum": [
          "DEFAULT_INCOME_ACCOUNT",
          "DEFAULT_EXPENSE_ACCOUNT",
          "DEFAULT_TRANSFER_SOURCE_ACCOUNT",
          "DEFAULT_TRANSFER_TARGET_ACCOUNT",
          "TRANSACTION_LIST_CHUNK_SIZE",
          "UPDATE_INVESTMENT_MARKET_VALUE",
          "DASHBOARD_SETTINGS",
          "STATISTICS_SETTINGS"
        ]
      },
      "UserPreferenceDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "key": {
            "allOf": [{ "$ref": "#/components/schemas/UserPreferenceProperty" }]
          },
          "value": { "type": "string" }
        },
        "required": ["createdAt", "updatedAt", "id", "userId", "key", "value"]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "nickname": { "type": "string" },
          "githubId": { "type": "string", "nullable": true },
          "auth0Id": { "type": "string", "nullable": true },
          "profileImageUrl": { "type": "string", "nullable": true },
          "roles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Role" }
          },
          "theme": { "allOf": [{ "$ref": "#/components/schemas/Theme" }] }
        }
      },
...
      "UpdateUserPreferenceDto": {
        "type": "object",
        "properties": {
          "key": {
            "allOf": [{ "$ref": "#/components/schemas/UserPreferenceProperty" }]
          },
          "value": { "type": "string" }
        },
        "required": ["key", "value"]
      },
...
    }
  }
}